JDP210701-5 - update user controller implementation
